functions/process-image/requirements.txt
google-cloud-storage
google-cloud-pubsub
Pillow
functions-framework